dataset/requirements.txt
httpx>=0.27.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
tabula-py>=2.9.0
tenacity>=8.2.0
rapidfuzz>=3.5.0
